requirements.txt
Flask
google-cloud-storage
google-generativeai
gunicorn
